support NMS scoreboard 1.17 (first try - part two)
I just realized that both packet classes
have "private final" in all fields...

and in JDK-16 that's impossible to hijack except with messing
with the security manager which is a pure evil thing to do.

So, basically I had to create NMS scoreboard
and then instantiate the NMS packets based on them.